+++ src/components/cymap/Hexagon/index.scss
@@ -1,0 +1,4 @@
+.map-content {
+  width: 100%;
+  height: 100%;
+}
