+++ src/components/echartComponents/FansRatio/index.jsx
@@ -1,0 +1,27 @@
+import './index.scss';
+import * as React from 'react';
+
+
+//粉丝占比
+class Fansration extends React.Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {};
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+  render() {
+    return (
+    <div className="Fansration">
+    </div>);
+  }
+}
+
+export default Fansration;
