+++ src/components/echartComponents/FansRatio/index.scss
@@ -1,0 +1,7 @@
+.Fansration {
+    width: 100%;
+    height: 100%;
+    background-color: #282c34;
+    background: url(../../../static/images/框1.png) center no-repeat;
+    background-size: cover;
+  }
