+++ src/components/echartComponents/Qyph/index.jsx
@@ -1,0 +1,27 @@
+import './index.scss';
+import * as React from 'react';
+
+
+//区域排行
+class index extends React.Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {};
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+  render() {
+    return (
+    <div className="Qyph">
+    </div>);
+  }
+}
+
+export default index;
