+++ src/components/echartComponents/Qyph/index.scss
@@ -1,0 +1,7 @@
+.Qyph {
+    width: 100%;
+    height: 100%;
+    background-color: #282c34;
+    background: url(../../../static/images/框3.png) center no-repeat;
+    background-size: cover;
+  }
