+++ src/components/echartComponents/Rcyt/index.jsx
@@ -1,0 +1,27 @@
+import './index.scss';
+import * as React from 'react';
+
+
+//热词云图
+class index extends React.Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {};
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+  render() {
+    return (
+    <div className="Rcyt">
+    </div>);
+  }
+}
+
+export default index;
