+++ src/components/echartComponents/Rcyt/index.scss
@@ -1,0 +1,7 @@
+.Rcyt {
+    width: 100%;
+    height: 100%;
+    background-color: #282c34;
+    background: url(../../../static/images/框5.png) center no-repeat;
+    background-size: cover;
+  }
