+++ src/components/echartComponents/Zfqk/index.jsx
@@ -1,0 +1,27 @@
+import './index.scss';
+import * as React from 'react';
+
+
+//涨粉情况
+class index extends React.Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {};
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+  render() {
+    return (
+    <div className="Zfqk">
+    </div>);
+  }
+}
+
+export default index;
