+++ src/components/echartComponents/Zfqk/index.scss
@@ -1,0 +1,7 @@
+.Zfqk {
+    width: 100%;
+    height: 100%;
+    background-color: #282c34;
+    background: url(../../../static/images/框2.png) center no-repeat;
+    background-size: cover;
+  }
