+++ src/components/echartComponents/Zyfx/index.jsx
@@ -1,0 +1,27 @@
+import './index.scss';
+import * as React from 'react';
+
+
+//职业分析
+class index extends React.Component {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {};
+  }
+
+  componentDidMount() {
+
+  }
+
+  componentWillUnmount() {
+
+  }
+
+  render() {
+    return (
+    <div className="Zyfx">
+    </div>);
+  }
+}
+
+export default index;
