+++ src/components/echartComponents/Zyfx/index.scss
@@ -1,0 +1,7 @@
+.Zyfx {
+    width: 100%;
+    height: 100%;
+    background-color: #282c34;
+    background: url(../../../static/images/框4.png) center no-repeat;
+    background-size: cover;
+  }
